second

+++ src/main/java/com/driver/Director.java
@@ -8,7 +8,7 @@ public class Director {
     public Director() {
     }
 
-    public Director(String name, int numberOfMovies, int imdbRating) {
+    public Director(String name, int numberOfMovies, double imdbRating) {
         this.name = name;
         this.numberOfMovies = numberOfMovies;
         this.imdbRating = imdbRating;
